:recycle:  refactor: Loader 컴포넌트 경로 organism으로 이동
