+++ src/components/refactor/atoms/index.ts
@@ -2,5 +2,4 @@ export { default as Button } from './Button';
 export { default as Input } from './Input';
 export { default as Img } from './Img';
 export { default as Icon } from './Icon';
-export { default as Loader } from './Loader';
 export { default as LottieLoader } from './LottieLoader';
