+++ src/components/refactor/organism/index.ts
@@ -1,0 +1,1 @@
+export { default as Loader } from './Loader';
